+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.0.4
+* Added much more sensors and created one sensor class (ProfiAirTouchSensor) for all sensor entities
+
+## v0.0.3
+* Changed namings and added translation keys
+* Added first sensors (4 temperature sensors)
+
 ## v0.0.2
 * Updated README.md and TODO.md